requirements.txt
fastapi
uvicorn[standard]

sentence-transformers
torch
transformers

numpy
pandas
scikit-learn

pyserini
faiss-cpu

pydantic
python-dotenv
PyYAML